requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf-yasg
django-cors-headers
setuptools
gunicorn
